test : category 정상 동작 및 예외 테스트
