+++ src/main/java/ddd/caffeine/ratrip/module/place/feign/kakao/model/FeignPlaceData.java
@@ -34,7 +34,7 @@ public class FeignPlaceData {
 			.telephone(this.phone)
 			.build();
 
-		place.setPlaceCategory(this.categoryGroupCode);
+		place.setCategoryByCode(this.categoryGroupCode);
 		place.setAddress(addressName);
 		place.setLocation(Double.parseDouble(y), Double.parseDouble(x));
